Final commit?
Documentation is needed for authors and what not. I'll be working
Monday, so I won't likely get to look at the code any more before we
submit. The SDL library is needed to compile and run the code. I can
send you guys an email with the compiled program if you want to test it.
let me know.